requirements.txt
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
trimesh==4.4.7
scipy==1.14.1
pytest==8.3.2
torch==2.4.1
torchvision==0.19.1
networkx==3.3
